requirements.txt
ipython==7.23.1
ipython-genutils==0.2.0
matplotlib==3.4.2
numpy==1.19.5
Pillow==8.2.0
tensorflow==2.3.0
pandas==1.2.2
Pillow==8.2.0
